Added Persistent Storage
Use the 'save' button to store the text in the text box to the currently selected date.
To access the data from another class, use the public method getPlannerData(date) where date is a String of the format "DD-MM-YYYY".
Returns an empty string if there was no data for the specified date.

Currently overhauling the barebones UI